Moved UI stuff in their own files
