+++ include/ui.h
@@ -1,0 +1,14 @@
+#ifndef WHEN_THE_FACTORY_UI_
+#define WHEN_THE_FACTORY_UI_
+
+#include <stdbool.h>
+#include "map.h"
+
+/* Selected tile, if any. */
+extern bool g_sel_tile_exists;
+extern TileCoords g_sel_tile_coords;
+
+void ui_select_tile(TileCoords tc);
+void refresh_selected_tile_ui(void);
+
+#endif /* WHEN_THE_FACTORY_UI_ */
